Speedup Histogram - Small Bug Fix
To accelerate the histogram, which some have complained causes the program to be laggy, I now only use 1/8th of the data coming off the camera to create the histogram. Seems fairly responsive on BT-MesoSPIM.

Also, @samcal06 found that if no channel settings are selected in the GUI, and we start up the software, setting the selected channel in the render_widgets of the camera_view controllers throws an error.